Move predicates to separate file
This places predicates affecting the gitOps controller next to each
other, in the same file, further improving readability of the
controller.
